new: MVC variant added + broken integration tests added

+++ src/main/kotlin/com/shirn/veggies/endpoint/VeggieType.kt
@@ -1,6 +1,6 @@
 package com.shirn.veggies.endpoint
 
-enum class VeggieType(val dbId: Int) {
+enum class VeggieType(val dbId: Long) {
     TOMATO(0),
     CUCUMBER(1),
     BANANA(2),
@@ -10,7 +10,7 @@ enum class VeggieType(val dbId: Int) {
         fun fromName(name: String): VeggieType {
             return entries.find { it.name == name } ?: throw IllegalStateException("Can't find a suitable veggie with $name")
         }
-        fun fromId(id: Int): VeggieType {
+        fun fromId(id: Long): VeggieType {
             return entries.find { it.dbId == id } ?: throw IllegalStateException("Can't find a suitable veggie with $id")
         }
     }
